remove unneeded class.
